separate out npm-specific ts code
