+++ lib/npm/exit0/darwin-x64-LE.node
