+++ lib/npm/exit0/win32-x64-LE.node
