+++ lib/npm/install.ts
