+++ lib/npm/worker.ts
