+++ lib/shared/common.ts
