+++ lib/shared/stdio_protocol.ts
